LayoutTests/media/video-seek-past-end-paused-expected.txt

Test that seeking paused video past it's duration time sets currentTime to duration and leaves video paused.

RUN(video.load())
EVENT(load)
RUN(video.currentTime = 500)
EVENT(timeupdate)
TEST(video.currentTime == video.duration) OK
END OF TEST
